refactor(matrix): split gateway helpers [OPE-393]
